+++ .github/workflows/bundle.yml
@@ -35,7 +35,7 @@ jobs:
         run: |
           git config user.name "github-actions[bot]"
           git config user.email "github-actions[bot]@users.noreply.github.com"
-          git add -f server.bundle.mjs cli.bundle.mjs hooks/session-extract.bundle.mjs hooks/session-snapshot.bundle.mjs hooks/session-db.bundle.mjs hooks/storage-paths.bundle.mjs hooks/security.bundle.mjs
+          git add -f server.bundle.mjs cli.bundle.mjs hooks/session-extract.bundle.mjs hooks/session-snapshot.bundle.mjs hooks/session-db.bundle.mjs hooks/security.bundle.mjs
           git diff --staged --quiet && echo "No changes" && exit 0
-          git commit -m "ci: update server.bundle.mjs, cli.bundle.mjs, session, storage & security hook bundles"
+          git commit -m "ci: update server.bundle.mjs, cli.bundle.mjs, session hook & security bundles"
           git push
