--- hooks/storage-paths.bundle.mjs
@@ -1,3 +1,0 @@
-import{accessSync as R,constants as V,mkdirSync as y}from"node:fs";import{dirname as D,isAbsolute as h,join as g,resolve as v}from"node:path";var o="CONTEXT_MODE_DIR",u="sessions",c="content",t=class extends Error{kind;path;overrideEnvVar;ignoredEnvVar;ignoredReason;constructor(n,e,i=o,a,E,d={}){super(E??x(n,e,d),{cause:a}),this.name="StorageDirectoryError",this.kind=n,this.path=e,this.overrideEnvVar=i,this.ignoredEnvVar=d.ignoredEnvVar,this.ignoredReason=d.ignoredReason}},s=new Map;function k(r,n,e){return new t(r,n,o,void 0,[`Invalid ${o} for context-mode ${r} directory: ${e}`,f()].join(`
-`))}function l(r){let n=process.env[o];if(n===void 0)return{kind:"unset"};let e=n.trim();if(!e)return{kind:"ignored-empty",ignoredEnvVar:o,ignoredReason:"empty"};if(!h(e))throw k(r,e,`${o} must be an absolute path.`);return{kind:"override",root:v(e)}}function m(r){return r.kind==="ignored-empty"?{ignoredEnvVar:r.ignoredEnvVar,ignoredReason:r.ignoredReason}:{}}function S(r,n){let e=l(r);return e.kind!=="override"?null:{kind:r,path:g(e.root,n),envVar:o,source:"override"}}function O(r,n,e){return{kind:r,path:v(n()),envVar:null,source:"default",...e}}function p(r){let n=l("session");return n.kind==="override"?{kind:"session",path:g(n.root,u),envVar:o,source:"override"}:O("session",r,m(n))}function $(r){let n=S("content",c);if(n)return n;let e=p(r);return{kind:"content",path:g(D(e.path),c),envVar:e.envVar,source:e.source,ignoredEnvVar:e.ignoredEnvVar,ignoredReason:e.ignoredReason}}function _(r){let n=S("stats",u);if(n)return n;let e=p(r);return{kind:"stats",path:e.path,envVar:e.envVar,source:e.source,ignoredEnvVar:e.ignoredEnvVar,ignoredReason:e.ignoredReason}}function C(r){return r.message}function N(r){return r.source==="override"&&r.envVar?`via ${r.envVar}`:r.ignoredEnvVar&&r.ignoredReason==="empty"?`default; ignored empty ${r.ignoredEnvVar}`:"default"}function j(){s.clear()}function B(r){let n=[r.kind,r.path,r.source,r.envVar??"",r.ignoredEnvVar??"",r.ignoredReason??""].join("\0"),e=s.get(n);if(e instanceof t)throw e;if(e===r.path)return e;try{return y(r.path,{recursive:!0}),R(r.path,V.W_OK),s.set(n,r.path),r.path}catch(i){let a=new t(r.kind,b(i)??r.path,o,i,void 0,{ignoredEnvVar:r.ignoredEnvVar,ignoredReason:r.ignoredReason});throw s.set(n,a),a}}function x(r,n,e={}){return[`context-mode ${r} directory is not writable: ${n}`,w(e),f()].filter(Boolean).join(`
-`)}function w(r){return r.ignoredEnvVar&&r.ignoredReason==="empty"?`Ignored empty ${r.ignoredEnvVar}; using adapter default.`:null}function f(){return`Set ${o} to a writable absolute path.`}function b(r){if(!r||typeof r!="object")return null;let n=r.path;return typeof n=="string"&&n.length>0?n:null}export{t as StorageDirectoryError,j as clearStorageDirectoryCheckCacheForTests,N as describeStorageDirectorySource,B as ensureWritableStorageDir,C as formatStorageDirectoryError,$ as resolveContentStorageDir,p as resolveSessionStorageDir,_ as resolveStatsStorageDir};
